Nội dung thay đổi mới

+++ Layout/src/main/java/com/example/layout/repository/PhanHoiRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import com.example.layout.entity.PhanHoi;
 
-public interface PhanHoiRepository extends JpaRepository<PhanHoi, Long> {
-	List<PhanHoi> findByKhachHangMaKhachHang(int maKhachHang); 
+public interface PhanHoiRepository extends JpaRepository<PhanHoi, Integer> {
+	List<PhanHoi> findByKhachHangMaKhachHang(Integer maKhachHang);
 }
